refactor: 공통되는 메뉴 로직을 컴포넌트로 분리
- navbar 로직이 중복되어서, LandingNavbar 컴포넌트로 분리해주었다.
